+++ Leomard/AppStorage/UserPreferences.swift
@@ -48,6 +48,7 @@ final class UserPreferences: ObservableObject {
     
     // MARK: - Navbar
     @AppStorage("navbar_show_letter_separators", store: .standard) var navbarShowLetterSeparators: Bool = false
+    @AppStorage("show_communities_instances", store: .standard) var showCommunitiesInstances: Bool = true
     
     // MARK: - Prefer Display Name
     @AppStorage("prefer_display_name_community_post", store: .standard) var preferDisplayNameCommunityPost: Bool = false